[splash-screen][Android] Scope kotlin objects (singletons) in separate package
