+++ android/versioned-abis/expoview-abi39_0_0/src/main/java/abi39_0_0/expo/modules/splashscreen/SplashScreenModule.kt
@@ -12,7 +12,7 @@ import abi39_0_0.org.unimodules.core.interfaces.ExpoMethod
 // Below import is added explicitly to provide a redirection from versioned code realm to unversioned code realm.
 // Without this import any `SplashScreen.methodName(...)` invocation on JS side ends up in versioned SplashScreen kotlin object that stores no information about the ExperienceActivity.
 // TODO (@bbarthec): add this import while performing Android versioning
-import expo.modules.splashscreen.SplashScreen
+import expo.modules.splashscreen.singletons.SplashScreen
 
 class SplashScreenModule(context: Context) : ExportedModule(context) {
   companion object {
